make :scoped and :history work together
do not complain when :scoped and :history are used together

keep track of the modules in use in Configuration.

This adds a uses? method which will return whether or not a given module is
used, e.g.

friendly_id_config.uses?(:scoped)

remove note about history module's incompatibility with the scope module

add back the index for the non-scoped case

handle scopes in the find_one and exists? overrides

when using :scoped friendly_ids aren't unique anymore, so we can't just use the
first match we find. Instead we now consider all ids in "with_old_friendly_id"
and use according conditions in "find_one" and "exists?" which will return the
proper object when used on the relation as adviced in the :scoped module
documentation, e.g.

@city.restaurants.find("joes-diner")