Fix slug creation in STI classes
Previously, if the child class had the finder modules loaded but the
parent class did not, then slugs could not be created because the slug
generator assumed friendly finders were available, but were not because
the scope is uses comes from the super class rather than the child
class.

The child class scope is not used in order to maintain slug uniqueness
across child classes.

See #516