Frienldy_id should not generate a slug when all of canidates is nil.
Previously `friendly_id_config.base` was checked on presence in `should_generate_new_friendly_id?`. And let's imagine
that we have multiple slug candidates, but all of them is nil. In this case we will have something like
`[nil, nil]` as base for slug and `presence?` for this array will be evalutated to true. But all of slug canidates is nil, right?
Booom! Exception during execution of `resolve_friendly_id_conflict`.