Defer loading of ActiveRecord to avoid config issues
If FriendlyId references `ActiveRecord::Base` while it is loading as
part of a Rails application, it triggers the `on_load` event for
ActiveRecord, which as part of the loading process sets the current
configuration values of ActiveRecord. If any initializers try to
influence that configuration (e.g. changing defaults in
`new_framework_defaults.rb` for an app that's been upgraded), these
changes will have no effect because the configuration has already been
set in place.

To avoid this, we need to avoid triggering the `on_load` event by
wrapping the reference to `ActiveRecord::Base` in an initializer
block. The simplest way of doing this is with a `Railtie`, which will
be automatically loaded if FriendlyId is required as part of Rails.

Fixes #823