Fix behaviour of to_param in callbacks
This ensures that we use the new value of a slug when calling
`to_param` within an ActiveRecord callback. It also seems to pass all
the other regression tests I've added, thus keeping the fixes working
for #259 and #148, and as a bonus, prevents getting a bunch of
deprecation warnings in Rails 5.1.