Persist sluggable_type to scope column when scope is not configured
This accounts for a subtle bug where the slugs can be duplicated if
the scope column is empty. The compound index
index_friendly_id_slugs_on_slug_and_sluggable_type_and_scope allows
duplicates if scope is empty.

This commit ensures that a fallback value, in this case `sluggable_type`,
is written to `scope` no matter what.