Pass all 4 parameters (message, model_name, query_field and id) to ActiveRecord::RecordNotFound.new when raising the exception. (#890)
In my Rails App, I had written a common rescue for `ActiveRecord::RecordNotFound` in `ApplicationController`.

When this exception is rescued, I form a common error message:

    rescue_from ActiveRecord::RecordNotFound, :with => :handle_404
    def handle_404
        @exception = exception
        @not_found_message = "The #{@exception.model.name.underscore.humanize} you are looking for was not found"
         render :not_found
    end

This worked fine when the model does not use `FriendlyId`. 

In order to make this work with models that use `FriendlyId`, I have passed all four parameters  (message, model_name, query_field and id) to `ActiveRecord::RecordNotFound.new` when it is being raised in finder_methods.rb.